Display an error if not the allowed values

This function calls the error() function to display an error if the variable's values are not valid.

assert_na(
  x,
  level = 1,
  msg_level = getOption("msgr.level"),
  msg_types = getOption("msgr.types"),
  log_path = getOption("msgr.log_path")
)

assert_in(
  x,
  values,
  level = 1,
  msg_level = getOption("msgr.level"),
  msg_types = getOption("msgr.types"),
  log_path = getOption("msgr.log_path")
)

assert_in_range(
  x,
  min = NULL,
  max = NULL,
  level = 1,
  msg_level = getOption("msgr.level"),
  msg_types = getOption("msgr.types"),
  log_path = getOption("msgr.log_path")
)

assert_char_length(
  x,
  n = NULL,
  n_min = NULL,
  n_max = NULL,
  level = 1,
  msg_level = getOption("msgr.level"),
  msg_types = getOption("msgr.types"),
  log_path = getOption("msgr.log_path")
)

Arguments

x
(any) The variable to check.

level
(integer, optional) The level of the message, from 1 to 10. Default: 1.

msg_level
(integer, optional) The maximum level of messages to output. Default: set in the option "msgr.level".

msg_types
(character, optional) The type to write or display. Must either NULL or one or more from "INFO", "WARNING" or "ERROR". Default: set in the option "msgr.types".

log_path
(character, optional) The file path to the text log file. If set to "", then no logs are written. Default: set in the option "msgr.log_path".

values
(character) The allowed values.

min
(numeric, optional) The minimum allowed value for a range. If NULL the minimum value is not checked. Default: NULL.

max
(numeric, optional) The maximum allowed value for a range. If NULL the maximum value is not checked. Default: NULL.

n
(integer, optional) The allowed length. If NULL the exact character length is not checked. Default: NULL.

n_min
(integer, optional) The allowed minimum length. If NULL the minimum character length is not checked. Default: NULL.

n_max
(integer, optional) The allowed maximum length. If NULL the maximum character length is not checked. Default: NULL.

Value

If assertion passes then TRUE is returned. This allows you to make multiple assertions separated by &.

Details

The following validations can be checked:

The following values can be checked:

assert_na(): All elements of 'x' must be NA.

assert_in(): All elements of 'x' must in specified values.

assert_in_range(): All elements of 'x' must be in the specified numeric range.

assert_char_length: All elements of 'x' must have valid character length. You can specify the exact length using n or the minimum and/or maximum length using n_min and n_max respectively.

Examples

if (FALSE) {

# No error
assert_na(NA)
assert_na(c(NA, NA, NA))
# Error
assert_na(1)
assert_na(c(1, NA, 3))

# No error
assert_in(1:3, 1:10)
assert_in(c("A", "B"), LETTERS)
# Error
assert_in(8:12, 1:10)
assert_in(c("A", "B"), letters)

# No error
assert_in_range(1:3, min = 1)
assert_in_range(1:3, min = 1, max = 10)
# Error
assert_in_range(1:3, min = 5)
assert_in_range(1:3, min = 1, max = 2)

# Error
assert_char_length(c("bob", "jan"), n = 3)
assert_char_length(c("bob", "jan"), n_min = 1)
assert_char_length(c("bob", "jan"), n_min = 1, n_max = 10)
# No error
assert_char_length(c("bob", "jane"), n = 3)
assert_char_length(c("bob", "jane"), n_min = 5)
assert_char_length(c("bob", "jane"), n_min = 1, n_max = 2)

}