Checks the variable values

This function returns TRUE if the variable values are correct and FALSE otherwise.

is_na(x)

is_in(x, values)

is_in_range(x, min = NULL, max = NULL)

has_char_length(x, n = NULL, n_min = NULL, n_max = NULL)

Arguments

x
(any) The object with elements to test

values
(any) The allowed (or not allowed) values.

min
(numeric, optional) The minimum allowed value for a range. If NULL the minimum value is not checked. Default: NULL.

max
(numeric, optional) The maximum allowed value for a range. If NULL the maximum value is not checked. Default: NULL.

n
(integer, optional) The allowed length. If NULL the exact character length is not checked. Default: NULL.

n_min
(integer, optional) The allowed minimum length. If NULL the minimum character length is not checked. Default: NULL.

n_max
(integer, optional) The allowed maximum length. If NULL the maximum character length is not checked. Default: NULL.

Value

TRUE if elements in x are in y, FALSE otherwise

Details

The following values can be checked:

is_na(): Elements of 'x' must be NA.

is_in(): Elements of 'x' must in specified values.

is_in_range(): Elements of 'x' must be in the specified numeric range.

has_char_length: Elements of 'x' must have valid character length. You can specify the exact length using n or the minimum and/or maximum length using n_min and n_max respectively.

Examples

# TRUE
is_na(NA)

#> [1] TRUE

is_na(c(NA, NA))

#> [1] TRUE TRUE

# FALSE
is_na(NULL)

#> [1] FALSE

is_na(1)

#> [1] FALSE

is_na("foo")

#> [1] FALSE

is_na(c(1, NA))

#> [1] FALSE  TRUE

# TRUE
is_in("a", letters)

#> [1] TRUE

is_in(c("a", "b", "c"), letters)

#> [1] TRUE TRUE TRUE

# FALSE
is_in(1, LETTERS)

#> [1] FALSE

is_in(1:2, LETTERS)

#> [1] FALSE FALSE

# TRUE
is_in_range(1, min = 0)

#> [1] TRUE

is_in_range(c(1.1, 2.2, 3.3), min = 0, max = 10)

#> [1] TRUE TRUE TRUE

# FALSE
is_in_range(0, min = 1)

#> [1] FALSE

is_in_range(c(1.1, 2.2, 3.3), min = 2, max = 3)

#> [1] FALSE  TRUE FALSE

# TRUE
has_char_length("", 0)

#> [1] TRUE

has_char_length(c("some", "text"), 4)

#> [1] TRUE TRUE

has_char_length(c("some", "text"), n_min = 1)

#> [1] TRUE TRUE

has_char_length(c("some", "text"), n_max = 10)

#> [1] TRUE TRUE

has_char_length(c("some", "text"), n_min = 1, n_max = 10)

#> [1] TRUE TRUE

# FALSE
has_char_length("sometext", 4)

#> [1] FALSE

has_char_length(c("different", "text"), n_min = 5)

#> [1]  TRUE FALSE

has_char_length(c("different", "text"), n_max = 5)

#> [1] FALSE  TRUE

has_char_length(c("different", "text"), n_min = 5, n_max = 8)

#> [1] FALSE FALSE