Check the variable type & class

This function returns TRUE if the variable type is correct and FALSE otherwise. Depending on the type, a number of other attributes can be checked simultaneously, such as length, using n, or size of rows and columns, using n_row and n_col respectively.

is_type(x, type, n = NULL)

is_class(x, class, n = NULL)

is_natural(x, n = NULL)

is_number(x, n = NULL)

is_factor(x, levels = NULL, n = NULL)

is_array(x, dims = NULL)

is_matrix(x, n_col = NULL, n_row = NULL)

is_data_frame(x, n_col = NULL, n_row = NULL)

Arguments

x
(any) The object to test.

type
(character) The allowed type.

n
(integer, optional) The allowed length. If NULL the length is not checked. Default: NULL.

class
(character) The allowed class.

levels
(character, optional) The allowed levels. If NULL the levels are not checked. Default: NULL.

dims
(integer, optional) The allowed dimension sizes. If NULL the dimensions are not checked. Default: NULL.

n_col
(integer, optional) The allowed number of columns. If NULL the columns are not checked. Default: NULL.

n_row
(integer, optional) The allowed number of rows. If NULL the rows are not checked. Default: NULL.

Value

TRUE if x is a valid URL, FALSE otherwise

Details

The following types can be checked:

is_type(): 'x"' must have type type and optionally length n.

is_class(): 'x"' must have class class and optionally length n.

is_natural(): 'x' must be a positive integer vector and optionally length n.

is_number(): 'x' must be a numeric vector and optionally length n.

is_factor(): 'x' must be a factor vector and optionally length n.

is_array(): 'x' must be an array and optionally have dimension size dims.

is_matrix(): 'x' must be a matrix and optionally have number of columns n_col and number of rows n_row.

is_data_frame(): 'x' must be a data.frame and optionally have number of columns n_col and number of rows n_row.

See also

Examples

# TRUE
is_type(1:3, "integer")

#> [1] TRUE

is_type(data.frame(a = 1:26, b = letters), "list")

#> [1] TRUE

is_type(1:3, "integer", n = 3)

#> [1] TRUE

# FALSE
is_type(data.frame(a = 1:26, b = letters), "data.frame")

#> [1] FALSE

is_type(1:3, "integer", n = 1)

#> [1] FALSE

# TRUE
is_class(1:3, "integer")

#> [1] TRUE

is_class(data.frame(a = 1:26, b = letters), "data.frame")

#> [1] TRUE

is_class(1:3, "integer", n = 3)

#> [1] TRUE

# FALSE
is_class(data.frame(a = 1:26, b = letters), "list")

#> [1] FALSE

is_class(1:3, "integer", n = 1)

#> [1] FALSE

# TRUE
is_natural(1:3)

#> [1] TRUE

is_natural(c(1.0, 2.0))

#> [1] TRUE

is_natural(1:3, n = 3)

#> [1] TRUE

# FALSE
is_natural(-1:-3)

#> [1] FALSE

is_natural(3.142)

#> [1] FALSE

is_natural(1:3, n = 1)

#> [1] FALSE

# TRUE
is_number(1:3)

#> [1] TRUE

is_number(c(1.2, 2.4))

#> [1] TRUE

is_number(1:3, n = 3)

#> [1] TRUE

# FALSE
is_number("text")

#> [1] FALSE

is_number(1:3, n = 1)

#> [1] FALSE

# TRUE
is_factor(factor(c("a", "b", "a")))

#> [1] TRUE

is_factor(factor(c("a", "b", "a")), n = 3)

#> [1] TRUE

is_factor(factor(c("a", "b", "a")), levels = c("a", "b"))

#> [1] TRUE

# FALSE
is_factor(1:3)

#> [1] FALSE

is_factor(factor(c("a", "b", "a")), n = 1)

#> [1] FALSE

is_factor(factor(c("a", "b", "a")), levels = c("A", "B"))

#> [1] FALSE

# TRUE
is_array(array(1:3))

#> [1] TRUE

is_array(array(1:12, dim = c(3, 4)))

#> [1] TRUE

is_array(array(1:12, dim = c(3, 4)), dims = c(3, 4))

#> [1] TRUE

# FALSE
is_array(1:3)

#> [1] FALSE

is_array(array(1:12, dim = c(3, 4)), dims = c(4, 3))

#> [1] FALSE

# TRUE
is_matrix(matrix(1:3))

#> [1] TRUE

is_matrix(matrix(1:12, nrow = 3, ncol = 4), n_col = 4)

#> [1] TRUE

is_matrix(matrix(1:12, nrow = 3, ncol = 4), n_row = 3)

#> [1] TRUE

# FALSE
is_matrix(1:3)

#> [1] FALSE

is_matrix(matrix(1:12, nrow = 3, ncol = 4), n_col = 3)

#> [1] FALSE

is_matrix(matrix(1:12, nrow = 3, ncol = 4), n_row = 4)

#> [1] FALSE

# TRUE
is_data_frame(data.frame(a = 1:26, b = letters))

#> [1] TRUE

is_data_frame(data.frame(a = 1:26, b = letters), n_col = 2)

#> [1] TRUE

is_data_frame(data.frame(a = 1:26, b = letters), n_row = 26)

#> [1] TRUE

# FALSE
is_data_frame(list(a = 1:26, b = letters))

#> [1] FALSE

is_data_frame(data.frame(a = 1:26, b = letters), n_col = 5)

#> [1] FALSE

is_data_frame(data.frame(a = 1:26, b = letters), n_row = 5)

#> [1] FALSE