Display a warning, and record in a log file, if a condition is true.

This function calls the warn() function to display a warning if the specified condition is true. If a message is not specified then a generic message is displayed.

warn_if(
  condition,
  ...,
  level = 1,
  msg_level = getOption("msgr.level"),
  msg_types = getOption("msgr.types"),
  log_path = getOption("msgr.log_path")
)

Arguments

condition
(boolean) The condition to check.

...
(strings) message to be displayed or written to file.

level
(integer, optional) The level of the message, from 1 to 10. Default: 1.

msg_level
(integer, optional) The maximum level of messages to output. Default: set in the option "msgr.level".

msg_types
(character, optional) The type to write or display. Must either NULL or one or more from "INFO", "WARNING" or "ERROR". Default: set in the option "msgr.types".

log_path
(string, optional) The file path to the text log file. If set to "", then no logs are written. Default: set in the option "msgr.log_path".

Value

A string is return invisibly containing the warning.

Examples

# Use warn_if() to create conditional timed warnings
warn_if(2 > 1, "Condition is true so this warning is shown")
warn_if(1 > 2, "Condition is false so this warning is not shown")

# As with warn() a level can be set
warn_if(2 > 1, "This level 2 warning is not shown by default", level = 2)

# Set default level in options to determine what is shown
options(msgr.level = 2)
warn_if(2 > 1, "This level 2 warning is shown now", level = 2)

# Set message types in options to determine what is shown
options(msgr.types = "ERROR")
warn_if(2 > 1, "This is warning, so will not be shown now")